refactor: 영화 상제 정보 팝업 관련 로직 moviePopup.ts로 분리
